+++ nucleus-support-tests/src/androidTest/java/nucleus/view/NucleusFragmentActivityTest.java
@@ -9,11 +9,6 @@ public class NucleusFragmentActivityTest extends BaseViewTest<NucleusFragmentAct
     }
 
     @Override
-    protected Class<? extends Presenter> getPresenterClass() {
-        return Presenter.class;
-    }
-
-    @Override
     protected Object getView() {
         return getActivity();
     }