+++ nucleus-support-v4/src/main/java/nucleus/view/NucleusSupportFragment.java
@@ -46,7 +46,7 @@ public class NucleusSupportFragment<PresenterType extends Presenter> extends Fra
      * The factory class used to create the presenter. Defaults to
      * {@link ReflectionPresenterFactory} to create the presenter
      * using a no arg constructor.
-     *
+     * <p/>
      * Subclasses can override this to provide presenters in other
      * ways, like via their dependency injector.
      */