+++ nucleus/src/androidTest/java/nucleus/presenter/RxPresenterTest.java
@@ -33,7 +33,8 @@ public class RxPresenterTest extends InstrumentationTestCase {
     }
 
     static PresenterFactory<TestPresenter> testPresenterFactory = new PresenterFactory<TestPresenter>() {
-        @Override public TestPresenter createPresenter() {
+        @Override
+        public TestPresenter createPresenter() {
             return new TestPresenter();
         }
     };