+++ nucleus/src/main/java/nucleus/view/NucleusLayout.java
@@ -7,9 +7,9 @@ import android.os.Parcelable;
 import android.util.AttributeSet;
 import android.widget.FrameLayout;
 
-import nucleus.presenter.Presenter;
 import nucleus.factory.PresenterFactory;
 import nucleus.factory.ReflectionPresenterFactory;
+import nucleus.presenter.Presenter;
 
 /**
  * This view is an example of how a view should control it's presenter.
@@ -62,33 +62,26 @@ public abstract class NucleusLayout<PresenterType extends Presenter> extends Fra
         helper.dropView();
     }
 
+
     // The following section can be copy & pasted into any View class, just update their description if needed.
 
     /**
-     * The factory class used to create the presenter. Defaults to
-     * {@link ReflectionPresenterFactory} to create the presenter
+     * The factory class used to create the presenter. Defaults to {@link ReflectionPresenterFactory} to create the presenter
      * using a no arg constructor.
-     *
-     * Subclasses can override this to provide presenters in other
-     * ways, like via their dependency injector.
+     * <p/>
+     * Subclasses can override this to provide presenters in other ways, like via their dependency injector.
      *
      * @return The {@link PresenterFactory} that can build a {@link Presenter}, or null.
      */
     public PresenterFactory<PresenterType> getPresenterFactory() {
-        Class<PresenterType> presenterClass = findPresenterClass(getClass());
-        return presenterClass == null ? null : new ReflectionPresenterFactory<>(presenterClass);
-    }
-
-    private Class<PresenterType> findPresenterClass(Class<?> viewClass) {
-        RequiresPresenter annotation = viewClass.getAnnotation(RequiresPresenter.class);
-        //noinspection unchecked
-        return annotation == null ? null : (Class<PresenterType>)annotation.value();
+        return ReflectionPresenterFactory.fromViewClass(getClass());
     }
 
     /**
      * Returns a current attached presenter.
      * This method is guaranteed to return a non-null value between
-     * onAttachedToWindow/onDetachedFromWindow calls.
+     * onResume/onPause and onAttachedToWindow/onDetachedFromWindow calls
+     * if the presenter factory returns a non-null value.
      *
      * @return a current attached presenter or null.
      */