+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### XX.05.2015, Version: 1.3.0
+
+* An ability to instantiate presenters with custom PresenterFactory, this allows
+  to put arguments into a presenter's constructor or to make an instance-specific
+  dependency injection.
+* NucleusAppCompatActivity
+
+
 ### 03.04.2015, Version: 1.1.2
 
 * Separate `PresenterHelper` class for easier View class creation.